reorganise and add some tests
