+++ test/cert/test_mac_command/common.cpp
@@ -1,0 +1,3 @@
+#include "common.h"
+
+TestServerState server_state;
